refactor: move collection facade hooks and selectors to dedicated collectionStore
