+++ src/renderer/components/mainWindow/bodyTabs/InputTabs/tabs/AuthorizationTab/AuthorizationTab.tsx
@@ -1,4 +1,4 @@
-import { selectRequest, useCollectionActions, useCollectionStore } from '@/state/appStore';
+import { selectRequest, useCollectionActions, useCollectionStore } from '@/state/collectionStore';
 import {
   AuthorizationInformation,
   AuthorizationType,
