+++ src/renderer/components/shared/settings/CollectionSettingsModal.tsx
@@ -22,7 +22,8 @@ import {
   useEnvironmentStore,
 } from '@/state/environmentStore';
 import { variableArrayToMap, variableMapToArray } from '@/state/helper/variableMappers';
-import { useAppStore, useCollectionStore } from '@/state/appStore';
+import { useAppStore } from '@/state/appStore';
+import { useCollectionStore } from '@/state/collectionStore';
 import { deepEqual } from '@/util/object-util';
 import { ClientCertificate } from 'shim/objects/collection';
 import { VariableObjectWithKey } from 'shim/objects/variables';
