+++ src/renderer/view/CollectionImport.tsx
@@ -5,7 +5,7 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { RendererEventService } from '@/services/event/renderer-event-service';
-import { useCollectionActions } from '@/state/appStore';
+import { useCollectionActions } from '@/state/collectionStore';
 import { Plus } from 'lucide-react';
 import {
   Dialog,
